+++ docs/src/changelog.md
@@ -11,7 +11,7 @@ Changelog](https://keepachangelog.com).
 
 ### Added
 
-- A `throw_on_error` argument to [`poll_loop()`](@ref) ([#9]).
+- A `throw` argument to [`poll_loop()`](@ref) ([#9]).
 - Support for some more options in [`Session`](@ref) ([#9]).
 - A new method for [`PKI.get_fingerprint_hash(::PKI.SshKey)`](@ref) to get a
   public key fingerprint straight from a [`PKI.SshKey`](@ref) ([#9]).
@@ -24,6 +24,9 @@ Changelog](https://keepachangelog.com).
 - [`authenticate()`](@ref) will now do host verification as well. This is
   critical for security so it is *strongly recommend* that all dependencies
   update to this release ([#9]).
+- All the `throw_on_*` arguments in the various `Session` and `SshChannel`
+  methods have been renamed `throw` for consistency with `Base` and the new
+  `throw` arguments in some auto-wrapped bindings ([#9]).
 
 ## [v0.3.0] - 2024-03-10